#3400 Remove unnecessary casts to long
Delete unnecessary tests. Update DoubleAnalyzer.

+++ processor/src/main/java/org/mapstruct/ap/internal/util/NativeTypes.java
@@ -284,7 +284,7 @@ public class NativeTypes {
 
         @Override
         public Class<?> getLiteral() {
-            return float.class;
+            return double.class;
         }
 
     }
